Specify --arch/--platform when compiling libraries
These arguments matter since the compiler can potentially emit different IL depending on arch.
For example, if the arch supports strlen intrinsic then ILOP_STRLEN instruction is used, otherwise a call to __strlen is generated instead.
Trying to use the ILOP_STRLEN instruction on an arch that doesn't actually support it will spit out an error when trying to actually generate the machine code.
A simple test that shows this exact issue is trying to compile a simple printf "Hello world" program for aarch64.

Based on git history, seems these arguments were passed back when Makefile was being used, but was removed unintentionally(?) when converting to CMake.