Fix addrof operator on arrays
Getting address of an array should just return the address of the array. Previous behaviour was to get the address of a variable that points to the array.

The type of addrof on array should probably be pointer to array that can implicitly decay into regular pointer, but I just made it be a pointer to the underlying array type.

I suspect that there is still some incorrectness when interacting with deref operator too, but this is good enough for most common case.